Move gPickupPeep* to PickupPeep.{cpp,h}
