+++ src/openrct2/drawing/PickupPeep.h
@@ -1,0 +1,28 @@
+/*****************************************************************************
+ * Copyright (c) 2014-2026 OpenRCT2 developers
+ *
+ * For a complete list of all authors, please refer to contributors.md
+ * Interested in contributing? Visit https://github.com/OpenRCT2/OpenRCT2
+ *
+ * OpenRCT2 is licensed under the GNU General Public License version 3.
+ *****************************************************************************/
+
+#pragma once
+
+#include <cstdint>
+
+struct ImageId;
+struct ZoomLevel;
+
+namespace OpenRCT2::Drawing
+{
+    struct RenderTarget;
+}
+
+extern ImageId gPickupPeepImage;
+extern int32_t gPickupPeepX;
+extern int32_t gPickupPeepY;
+extern ZoomLevel gPickupPeepZoom;
+
+void GfxInvalidatePickedUpPeep();
+void GfxDrawPickedUpPeep(OpenRCT2::Drawing::RenderTarget& rt);
